// FetchToSubgrid/@types/enums.ts
export enum AttributeType {
  TwoOptions = 0,
  Customer = 1,
  DateTime = 2,
  DecimalNumber = 3,
  FloatNumber = 4,
  Number = 5,
  Lookup = 6,
  Money = 8,
  Owner = 9,
  PickList = 11,
  MultiselectPickList = 17,
  Status = 13,
  PrimaryKey = 15,
  StateCode = 12,
}

export enum WholeNumberType {
  Number = '0',
  Duration = '1',
  TimeZone = '2',
  Language = '3',
}